+++ regreg/smooth/cox_utils.pyx
@@ -2,9 +2,9 @@ import warnings
 import numpy as np, cython
 cimport numpy as cnp
 
-DTYPE_float = np.float
+DTYPE_float = float
 ctypedef cnp.float_t DTYPE_float_t
-DTYPE_int = np.int
+DTYPE_int = int
 ctypedef cnp.int_t DTYPE_int_t
 ctypedef cnp.intp_t DTYPE_intp_t
